+++ pkg/cloud/fakes.go
@@ -37,7 +37,7 @@ func (f *FakeCloudProvider) DeleteDisk(volumeID string) (bool, error) {
 	return true, nil
 }
 
-func (f *FakeCloudProvider) GetVolumeByNameAndSize(name string, size int) (string, error) {
+func (f *FakeCloudProvider) GetVolumeByNameAndSize(name string, size int64) (string, error) {
 	var disks []*fakeDisk
 	for _, disk := range f.disks {
 		for key, value := range disk.options.Tags {
